Validate placeholders once for whole schema
